pyproject.toml
[project]
name = "miroflow"
version = "1.6.0"
description = "MiroFlow: Performance-first agent framework that makes any model better"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11,<3.14"
authors = [
    { name = "MiromindAI" },
]
keywords = ["agent", "llm", "mcp", "benchmark", "miroflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    # === Core ===
    "hydra-core>=1.3.2",
    "mcp>=1.12.2",
    "fastmcp>=2.10.6",
    "rich>=14.1.0",
    "tenacity>=8.2.3,<9.0.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.1.1",
    "jinja2",
    "json5>=0.12.0",
    "requests>=2.32.4",
    "aiohttp>=3.12.15",
    "filetype>=1.2.0",
    "Pillow",
    "fire>=0.7.0",
    # === LLM Providers ===
    "openai==1.78.1",
    "tiktoken>=0.9.0",
    "anthropic>=0.60.0",
    "google-genai>=1.28.0",
    # === Web ===
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # === Benchmark ===
    "datasets<4.0.0",
    "pandas>=2.3.0",
    # === File Processing ===
    "mammoth>=1.6.0",
    "pdfminer.six>=20221105",
    "python-pptx>=0.6.21",
    "openpyxl>=3.1.5",
    "markitdown>=0.1.0",
    "markitdown-mcp>=0.0.1a3",
    "markdownify>=0.11.0",
    "beautifulsoup4>=4.12.0",
    # === Sandbox ===
    "e2b-code-interpreter>=1.2.1",
    # === Other ===
    "tencentcloud-sdk-python>=3.0.0",
    "mutagen>=1.47.0",
    "wikipedia>=1.4.0",
    "pyzmq",
    # === Docs ===
    "mkdocs>=1.6.0,<2.0.0",
    "mkdocs-material>=9.5.0,<9.6.0",
    "pymdown-extensions>=10.0",
]

[project.urls]
Homepage = "https://github.com/MiroMindAI/miroflow"
Documentation = "https://miromindai.github.io/miroflow/"

[project.optional-dependencies]
dev = [
    "ruff==0.8.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "pre-commit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["miroflow"]
